working draft, see desc.
- just like Konva.js docs recommended, don't use sceneFunc for any of side-effects, then I though I could manipulate DataImage using Konva.Image as some sort of broker, so I've tried it and it just worked !
