+++ src/konva-components/index.js
@@ -3,10 +3,12 @@ import { Layer } from 'konva/lib/Layer';
 import { Group } from 'konva/lib/Group';
 import { Node } from 'konva/lib/Node';
 import { Shape } from 'konva/lib/Shape';
+import { Image } from 'konva/lib/shapes/Image';
 export { 
     Stage, 
     Layer, 
     Group, 
     Node, 
     Shape,
+    Image as KonvaImage
 }
